extract pure functions to separate modules
